Fix an underflow that causes panics in timekeeper
If the value internal to the Instant is smaller than the server tick (e.g. on Windows clients, when the client OS uptime is smaller than the game server uptime), then the calculation can underflow.

Instead, do the subtraction in u64 ticks, and don't try to calculate Duration objects that represent times before the client started.